requirement.txt
django
psycopg2
djangorestframework
djangorestframework-simplejwt
django-filter